+++ src/main/kotlin/camp/nextstep/lotto/raffle/LottoResultMatcher.kt
@@ -1,4 +1,4 @@
-package camp.nextstep.lotto
+package camp.nextstep.lotto.raffle
 
 import camp.nextstep.lotto.ticket.LottoTicket
